chore: auth 버튼 디렉토리 구조 변경
